requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
transformers==4.35.2
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.1
jinja2==3.1.2
google-generativeai==0.3.2
